+++ CHANGES.md
@@ -24,12 +24,14 @@ Changes in 5.9
 - HTML add .br-after span to Course Periods + use DBGetOne() in Side.php
 - jQuery already loaded on first-login page, remove it in FirstLogin.fnc.php
 - CSS WPadmin pointer cursor for radio & checkbox input in stylesheet.css
+- Move REMOVE_ACCESS_USERNAME_PREFIX_ADD from PROGRAM_CONFIG (per school) to CONFIG (all schools, 0) in RemoveAccess.php, rosariosis.sql & Update.fnc.php
 
 Changes in 5.8.1
 ----------------
 - Fix Next Year widget: Handle "Retain" case: value is '0' in Widgets.fnc.php
 - Fix Search: Use POST for Public Pages plugin compatibility in Courses.php
 - Fix MarkDown preview + bold in warehouse.js & Inputs.php
+- Fix saving REMOVE_ACCESS_USERNAME_PREFIX_ADD on Grant in RemoveAccess.php
 
 Changes in 5.8
 --------------